--- src/datatypes.ts
@@ -1,15 +1,0 @@
-/**
- * @enum {string} Aliases for DataTypes used in a {@link ModelColumn} definition.
- */
-export enum DataTypes {
-	INTEGER = "integer",
-	INT = "integer",
-	TEXT = "text",
-	STRING = "text",
-	VARCHAR = "text",
-	CHAR = "text",
-	NUMBER = "real",
-	NUMERIC = "real",
-	REAL = "real",
-	BLOB = "blob",
-}
