+++ src/index.ts
@@ -1,4 +1,3 @@
 export * from "./database.js";
 export * from "./model.js";
-export * from "./datatypes.js";
 export * from "./queryBuilder.js";
